+++ tmpl/default.php
@@ -3,7 +3,7 @@
  * @package     Joomla.Site
  * @subpackage  mod_lupo_categories
  *
- * @copyright   Copyright (C) 2014 databauer / Stefan Bauer, Inc. All rights reserved.
+ * @copyright   Copyright (C) 2015 databauer / Stefan Bauer, Inc. All rights reserved.
  * @license     GNU General Public License version 2 or later
  */
 
@@ -25,8 +25,28 @@ defined('_JEXEC') or die;
 	<?php echo $module->content;?>
     <ul class="uk-list uk-list-space">
     <?php
-    foreach($list as $category){?>
-        <li><a href="<?=$category['link']?>"><?=$category['title']?></a> <div class="uk-badge"><?=$category['number']?></div></li>
-    <?php } ?>
+    if($params->get('show_categories')) {
+        foreach ($list['categories'] as $category) {
+            ?>
+            <li><a href="<?= $category['link'] ?>"><?= $category['title'] ?></a>
+
+                <div class="uk-badge uk-float-right"><?= $category['number'] ?></div>
+            </li>
+        <?php }
+    }
+
+    if($params->get('show_categories') && $params->get('show_agecategories')) {
+        ?><hr /><?php
+    }
+
+    if($params->get('show_agecategories')) {
+        foreach ($list['agecategories'] as $category) {
+            ?>
+            <li><a href="<?= $category['link'] ?>"><?= $category['title'] ?></a>
+
+                <div class="uk-badge uk-float-right"><?= $category['number'] ?></div>
+            </li>
+        <?php }
+    }?>
     </ul>
 </div>