+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 2.0.0 (Januar 27, 2015)
+
+  - Display of age-categories (optional)
+
+### 1.1.0 (Januar 26, 2015)
+
+  - Badge with number of toys aligned right
+
 ### 1.0.1 (October 15, 2014)
 
   - Install manifest updated to allow updating module